pyproject.toml
[project]
name = "chess-pgn-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "Charlie",email = "yilongchen007@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-chess>=1.999",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
